% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_motif_p_4n.R
\name{conf.motifs.p}
\alias{conf.motifs.p}
\title{Random networks confidence interval with paralelization}
\usage{
conf.motifs.p(
  pos,
  neg,
  network_name,
  num_random_networks,
  out,
  cores,
  square.motifs = FALSE
)
}
\arguments{
\item{pos}{Positive interactions. Input consists on a three colums data frame.}

\item{neg}{Negative interactions. Input consists on a three colums data frame.}

\item{network_name}{Network name which will be included on the output table.}

\item{num_random_networks}{Number of random networks generated for calculating the confidence
intervals for each motiv.}

\item{out}{Out format. Choose between "count" (number of motifs) or "normalized"
(number of motifs divided by the square of the node number).}

\item{cores}{Number of cores for paralelization. If missing, the number of available
cores will be autimatically detected and the parameter will be stablished as n°cores - 1.}

\item{square.motifs}{TRUE or FALSE to include or not the four node motifs.}
}
\value{
List of...
}
\description{
For a given trait, this function ...
}
\examples{

data(pos)
data(neg)

conf.motifs.p(pos = pos, neg = neg, network_name = "test_network",
num_random_networks = 10, out = "count", cores = 2, square.motifs = TRUE)

}
