% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifs}
\alias{motifs}
\title{Extract motifs from co-occurrence/co-exclusion networks.}
\usage{
motifs(pos, neg, network_name, out, square.motifs = FALSE)
}
\arguments{
\item{pos}{Positive interactions. Input consists on a three colums data frame.}

\item{neg}{Negative interactions. Input consists on a three colums data frame.}

\item{network_name}{Network name which will be included on the output table.}

\item{out}{Out format. Choose between "count" (number of motifs) or "normalized"
(number of motifs divided by the square of the node number).}

\item{square.motifs}{TRUE or FALSE to include or not the four node motifs.}
}
\value{
A list of two data frames...
}
\description{
Quantify 3 and square 4 nodes mofifs from co-occurrence and co-exclusions networks.
}
\examples{

data(pos)
data(neg)

motifs(pos = pos, neg = neg, network_name = "test_network", out = "count")
motifs(pos = pos, neg = neg, network_name = "test_network", out = "normalized", square.motifs = TRUE)

}
