% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_motif_p.R
\name{conf.motifs.p}
\alias{conf.motifs.p}
\title{Random networks confidence interval with paralelization}
\usage{
conf.motifs.p(pos, neg, network_name, num_random_networks, out, cores)
}
\arguments{
\item{pos}{Positive interactions. Input consists on a three colums data frame}

\item{neg}{}

\item{network_name}{Network name (character)}

\item{num_random_networks}{}

\item{out}{}

\item{cores}{}
}
\value{

}
\description{
For a given trait, this function ...
}
\examples{

data(pos)
data(neg)

conf.motifs(pos = pos, neg = neg, network_name = "test_network", num_random_networks = 100,
out = "count", cores = 4)

conf.motifs(pos = pos, neg = neg, network_name = "test_network", num_random_networks = 100,
out = "normalized", cores = 4)

}
