% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_motifs.R
\name{conf.motifs}
\alias{conf.motifs}
\title{Random networks confidence interval}
\usage{
conf.motifs(pos, neg, network_name, num_random_networks, out)
}
\arguments{
\item{pos}{Positive interactions. Input consists on a three colums data frame}

\item{neg}{}

\item{network_name}{Network name (character)}

\item{num_random_networks}{}

\item{out}{}
}
\value{

}
\description{
For a given trait, this function ...
}
\examples{

data(pos)
data(neg)

conf.motifs(pos = pos, neg = neg, network_name = "test_network", num_random_networks = 100,
out = "count")

conf.motifs(pos = pos, neg = neg, network_name = "test_network", num_random_networks = 100,
out = "normalized")

}
