% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifs}
\alias{motifs}
\title{Extract motifs from co-occurrence/co-exclusion networks.}
\usage{
motifs(pos, neg, network_name, out)
}
\arguments{
\item{pos}{Positive interactions. Input consists on a three colums data frame}

\item{neg}{}

\item{network_name}{}

\item{out}{}
}
\value{
A list of two data frames...
}
\description{
For a given trait, this function ...
}
\examples{

data(pos)
data(neg)

motifs(pos = pos, neg = neg, network_name = "test_network", out = "count")
motifs(pos = pos, neg = neg, network_name = "test_network", out = "normalized")

}
